+++ packages/controls/src/index.js
@@ -1,6 +1,0 @@
-export { default as Radio, RadioGroup } from "./radio"
-export { default as Checkbox } from "./checkbox"
-export * from "./switch"
-export { default as Button } from "./button"
-export * from "./select"
-export * from "./input"