+++ packages/controls/src/select/select-provider.tsx
@@ -1,36 +1,36 @@
-import { createContext, useReducer, useRef } from "react"
-import { Box } from "@rent_avail/core"
+import { createContext, PropsWithChildren, useReducer, useRef } from "react"
+import { Box, SXObject } from "@rent_avail/core"
 import { noop } from "@rent_avail/utils"
 
-/**
- * Select: Input that let's the user choose from a list of pre-defined values.
- *
- * Selects should be used to help a user choose from a list of 5 or more values. The
- * component shows the currently select value an allows the user to open a menu to change
- * their selection. Select components can be optionally searchable. To view helpful
- * examples to go https://design.avail.co/elements/controls#select
- *
- * @example
- *
- * <Select>
- *   <SelectInput label="State" />
- *   <SelectList>
- *     <SelectItem value="IL" label="Illinois" >Illinois</SelectItem>
- *   </SelectList>
- * </Select>
- */
+interface SelectProps {
+  multi?: boolean
+  defaultValue?: any[]
+  id: string
+  onSelect?: Function
+  sx?: SXObject
+}
 
 export const SelectContext = createContext(null)
 
-export const SelectActions = {
-  openList: "OPEN_LIST",
-  closeList: "CLOSE_LIST",
-  updateSearchValue: "UPDATE_SEARCH_VALUE",
-  selectValue: "SELECT_VALUE",
-  removeToken: "REMOVE_TOKEN",
+export enum SelectActions {
+  openList = "OPEN_LIST",
+  closeList = "CLOSE_LIST",
+  updateSearchValue = "UPDATE_SEARCH_VALUE",
+  selectValue = "SELECT_VALUE",
+  removeToken = "REMOVE_TOKEN",
 }
 
-function selectReducer(state, action) {
+type SelectState = {
+  open: boolean
+  multi: boolean
+  searchValue: string
+  tokens: any[]
+  id: string
+}
+
+type SelectAction = { type: SelectActions; payload: any }
+
+function selectReducer(state: SelectState, action: SelectAction) {
   switch (action.type) {
     case SelectActions.openList:
       return { ...state, open: true }
@@ -55,6 +55,24 @@ function selectReducer(state, action) {
   }
 }
 
+/**
+ * Select: Input that let's the user choose from a list of pre-defined values.
+ *
+ * Selects should be used to help a user choose from a list of 5 or more values. The
+ * component shows the currently select value an allows the user to open a menu to change
+ * their selection. Select components can be optionally searchable. To view helpful
+ * examples to go https://design.avail.co/elements/controls#select
+ *
+ * @example
+ *
+ * <Select>
+ *   <SelectInput label="State" />
+ *   <SelectList>
+ *     <SelectItem value="IL" label="Illinois" >Illinois</SelectItem>
+ *   </SelectList>
+ * </Select>
+ */
+
 export function Select({
   children,
   multi = false,
@@ -62,7 +80,7 @@ export function Select({
   id,
   sx,
   defaultValue = [],
-}) {
+}: PropsWithChildren<SelectProps>) {
   const inputRef = useRef(null)
   const listRef = useRef(null)
   const [state, dispatch] = useReducer(selectReducer, {