+++ packages/hooks/__tests__/hooks.tsx
@@ -1,0 +1,5 @@
+describe("Hooks", () => {
+  it("Tests Hooks", () => {
+    expect(true).toBe(true)
+  })
+})