+++ packages/popover/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.3.16](https://github.com/rentalutions/elements/compare/@rent_avail/popover@0.3.11...@rent_avail/popover@0.3.16) (2023-08-29)
+## [0.3.16](https://github.com/rentalutions/elements/compare/@rent_avail/popover@0.3.10...@rent_avail/popover@0.3.16) (2023-03-15)
 
 **Note:** Version bump only for package @rent_avail/popover