+++ packages/tag/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.23](https://github.com/rentalutions/elements/compare/@rent_avail/tag@0.1.21...@rent_avail/tag@0.1.23) (2023-08-29)
+## [0.1.23](https://github.com/rentalutions/elements/compare/@rent_avail/tag@0.1.21...@rent_avail/tag@0.1.23) (2023-03-15)
 
 **Note:** Version bump only for package @rent_avail/tag