+++ packages/toast/CHANGELOG.md
@@ -3,9 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.7](https://github.com/rentalutions/elements/compare/@rent_avail/toast@0.1.5...@rent_avail/toast@0.1.7) (2023-08-29)
+# [0.2.0](https://github.com/rentalutions/elements/compare/@rent_avail/toast@0.1.5...@rent_avail/toast@0.2.0) (2023-03-15)
 
-**Note:** Version bump only for package @rent_avail/toast
+### Features
+
+- **controls:** upgrade to typescript ([460f664](https://github.com/rentalutions/elements/commit/460f6642822e6586a9a54fec26fb2c61a0dd7d35))
 
 ## [0.1.6](https://github.com/rentalutions/elements/compare/@rent_avail/toast@0.1.5...@rent_avail/toast@0.1.6) (2022-11-09)