+++ packages/utils/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.5.3](https://github.com/rentalutions/elements/compare/@rent_avail/utils@0.5.1...@rent_avail/utils@0.5.3) (2023-08-29)
+## [0.5.3](https://github.com/rentalutions/elements/compare/@rent_avail/utils@0.5.1...@rent_avail/utils@0.5.3) (2023-03-15)
 
 **Note:** Version bump only for package @rent_avail/utils